refactor(helloworld): extract SayHello action to action.go
Move SayHelloActionName and SayHello from worker.go to a dedicated
action.go file, matching the standard WorkerBase file layout.
