+++ umh-core/pkg/fsmv2/workers/example/helloworld/action.go
@@ -1,0 +1,36 @@
+// Copyright 2025 UMH Systems GmbH
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package hello_world
+
+import "context"
+
+// SayHelloActionName is the name used for logging and metrics.
+const SayHelloActionName = "say_hello"
+
+// SayHello implements the say-hello action logic.
+// Idempotent: skips if hello was already said.
+func SayHello(_ context.Context, d *HelloworldDependencies) error {
+	logger := d.ActionLogger(SayHelloActionName)
+
+	if d.HasSaidHello() {
+		logger.Debug("already_said_hello")
+		return nil
+	}
+
+	logger.Info("hello_world")
+	d.SetHelloSaid(true)
+
+	return nil
+}
